Draft cluster snakefile stuff.
The "Snakefile_cluster_iteration" file manages the list of
widths, and can iterate over jobs on the cluster and then
do a (trivial) aggregation step.

The queuing_config.yaml file is a short cluster config that
removes the need to type the cluster stuff into the command
line.  It conventionally goes in
$HOME/.config/snakemake/queuing/config.yaml, and is accessed
by "snakemake <other-args> --profile=queuing".

The Snakefile has mpi stuff for the trivial aggregation rule
because the cluster-ness of it applies to all the rules.
This is clumsy, but not wrong.